Switch to using block timestamps from API.
